+++ apps/agentstack-ui/src/modules/platform-context/api/keys.ts
@@ -12,6 +12,4 @@ export const contextKeys = {
   histories: () => [...contextKeys.all(), 'history'] as const,
   history: ({ contextId, query = {} }: ListContextHistoryParams) =>
     [...contextKeys.histories(), contextId, query] as const,
-  historyInitial: ({ contextId, query = {} }: ListContextHistoryParams) =>
-    [...contextKeys.history({ contextId }), 'initial', query] as const,
 };
