+++ apps/agentstack-ui/src/modules/runs/chat/ChatView.tsx
@@ -4,13 +4,15 @@
  */
 
 import { extractServiceExtensionDemands, formExtension } from 'agentstack-sdk';
-import { useMemo } from 'react';
+import { useCallback, useMemo } from 'react';
 
 import { getAgentExtensions } from '#api/utils.ts';
 import { MainContent } from '#components/layouts/MainContent.tsx';
 import type { Agent } from '#modules/agents/api/types.ts';
 import { AgentDetailPanel } from '#modules/agents/components/detail/AgentDetailPanel.tsx';
+import { usePlatformContext } from '#modules/platform-context/contexts/index.ts';
 import { SourcesPanel } from '#modules/sources/components/SourcesPanel.tsx';
+import { routes } from '#utils/router.ts';
 
 import { FormRenderView } from '../components/FormRenderView';
 import { RunLandingView } from '../components/RunLandingView';
@@ -35,6 +37,7 @@ export function ChatView({ agent }: Props) {
 
 function Chat() {
   const { isPending, agent, hasMessages } = useAgentRun();
+  const { contextId } = usePlatformContext();
 
   useSyncRunStateWithRoute();
 
@@ -46,6 +49,19 @@ function Chat() {
     return formRender ?? undefined;
   }, [agent]);
 
+  const handleMessageSent = useCallback(() => {
+    if (contextId) {
+      window.history.pushState(
+        null,
+        '',
+        routes.agentRun({
+          providerId: agent.provider.id,
+          contextId,
+        }),
+      );
+    }
+  }, [agent, contextId]);
+
   const isLanding = !isPending && !hasMessages;
 
   return (
@@ -53,9 +69,9 @@ function Chat() {
       <MainContent spacing="md" scrollable={isLanding}>
         {isLanding ? (
           formRender ? (
-            <FormRenderView formRender={formRender} />
+            <FormRenderView formRender={formRender} onMessageSent={handleMessageSent} />
           ) : (
-            <RunLandingView />
+            <RunLandingView onMessageSent={handleMessageSent} />
           )
         ) : (
           <ChatMessagesView />
